+++ tests/PhpSpreadsheetTests/Reader/Xlsx/URLImageTest.php
@@ -26,7 +26,7 @@ class URLImageTest extends TestCase
             self::assertInstanceOf(Drawing::class, $drawing);
             // Check if the source is a URL or a file path
             self::assertTrue($drawing->getIsURL());
-            self::assertSame('https://www.globalipmanager.com/DataFiles/Pics/20/Berniaga.comahp2.jpg', $drawing->getPath());
+            self::assertSame('https://phpspreadsheet.readthedocs.io/en/latest/topics/images/01-03-filter-icon-1.png', $drawing->getPath());
             $imageContents = file_get_contents($drawing->getPath());
             self::assertNotFalse($imageContents);
             $filePath = tempnam(sys_get_temp_dir(), 'Drawing');
@@ -36,7 +36,7 @@ class URLImageTest extends TestCase
             unlink($filePath);
             self::assertNotFalse($mimeType);
             $extension = File::mime2ext($mimeType);
-            self::assertSame('jpeg', $extension);
+            self::assertSame('png', $extension);
         }
     }
 }